auth_service/requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn>=0.27.0,<0.30.0
sqlalchemy==2.0.29
asyncpg>=0.29.0,<0.30.0
redis>=5.0.0,<6.0.0
alembic>=1.11.0,<1.14.0
pydantic>=2.6.0,<3.0.0
pydantic-settings==2.10.0
email-validator>=2.1.0,<3.0.0
passlib>=1.7.4,<1.8.0
cryptography>=42.0.0,<43.0.0
slowapi==0.1.9
pyotp>=2.9.0,<3.0.0
python-jose==3.5.0
psycopg2-binary==2.9.10
qrcode[pil]>=7.3.1,<7.5.0
structlog>=24.1.0,<25.0.0
colorlog>=6.8.0,<6.9.0
httpx>=0.27.0,<0.28.0
requests>=2.31.0,<2.33.0
pytest>=8.1.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
pytest-cov>=5.0.0,<6.0.0
coverage>=7.4.0,<8.0.0
python-dotenv>=1.0.0,<1.1.0
